Eliminate systems based on difference between coefficients of equations.
This is an extension of the technique to look for equations for which the vectors of coefficients are at hamming  distance 1. Here we look for two equations for which the sign of the difference is the same for all coefficients, and for which not all coefficients of the difference are zero.